Check that the data dir exists
If you start up Redpanda point to a data dir that doesn't exist, the
error is fairly obscure: a std::system_error mentioning failed to open
but no indication of _what_ we were trying to open.

This has tripped up several people, including externally so let's
just fix it. We apply a check that the directory exists immediately
before creating the pidfile, which is the place that will fail first
in practice if the directory does not exist.

Pidfiles are enabled by default, but are disabled in some places, e.g.,
our fixture tests: if we *don't* create a pidfile we do not require
the directory to exist, it will be created automatically, so we don't
check existence in that case.