r/offset_translator: remove unsafe bootstrap code
When the corresponding kvstore state is not found, we should recover offset_translator
state from the log, not use configuration_manager state - the latter is incorrect because
there are other batch types contributing to offset delta, not just configuration batches.
This code is a vestige from the time when the separate offset_translator component was
just introduced and its kvstore state needed to be bootstrapped from configuration_manager
state.