ssx/work_queue: support recursive tasks
There are situations where we need to support recursive tasks for the
transform subsystem, and in order to do that we can't keep a single
variable of the tail, as it's possible for recursive tasks to try to
tail off a single future. Fix this by tracking the pending work in an
explicit data structure.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>
(cherry picked from commit 5ec499d7313b8156d9a799ec5623e27731a257a1)