c/log_eviction_stm: do not request snapshot if already progressed
It is perfectly possible that the install snapshot request will reach
the follower right before the `log_eviction_stm` asks for the snapshot
creation. In this case an `stm_manager` would do the check and throw an
exception informing that the snapshot can not be taken. In order to
handle the situation gracefully added a check in log eviction stm to
skip taking snapshot if start offset already progressed.

Fixes: #14220

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 443385160cbcc1e112e0d4ea64fd6c8c94827538)