tx_migration: avoid ping pong of requests between brokers
When leader table is stale (say at startup or during failures), the
current code can result in a ping pong of requests between two brokers
in a tight loop.

Example

- tx_migration_replicate dispatched from node 1 to node 2 (because node 1
thinks node 2 is the leader)
- node 2 dispatches the request back to node 1 because it thinks node 1
is the leader.

Until leadership stabilizes this results in a huge pile up of requests
which manifested as an oversized allocation.

Don't think a router is the right choice in the handler as the handler
is supposed to process the request locally. If it returns an error, it
is propagated to the source router which dispatches to the correct
leader. This also breaks the ping pong loop as the source router has
sleeps induced for retry backoff.

(cherry picked from commit a174ba25867fd7bdb40998e7e496af2904de78dc)