topic_recovery_test: adjust wait condition for finished uploads
Today the test waits for {the delta between cloud and local storage to
be steady for some time and under some size}, to denote quiesced state.

This meant that that even if we were no longer uploading to cloud, if
local storage were to be truncated continually, the test would not
consider the state quiesced, and then fail.

Instead it now waits for {the size in cloud to be steady for some time
and the delta between cloud and local storage to be under some size}.

Fixes #13406