consensus: aggregate per-partition metrics
When metrics aggregation is turned on, we want to aggregate away the
partition labels on most metrics: this wasn't occurring in the
consensus object for two metrics: leader_for and reconfiguration
changes in progress.

This change enable aggregation on the partition label for these metrics
if aggregation is turned on.

Issue #15811.
Issue redpanda-data/core-internal#677.