build: add include pattern to vlib helper
Adds <module>/include path to includes directory when linking against a
v::lib library, which is intended to be the pattern for exposing
public headers from a module.

The default include path of the root of the tree is also removed because
it is currently handled in the root cmake file and will later be
removed.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>