storage: Add 'offset_range_size' method
Add method that returns the on-disk size of the offset range. The method
uses segment indexes to locate the approx. location of the begining and
the end of the offset range and then scans segments until it finds the
precise location. The method only returns result if both offsets are
present in the log.