tests: Use multiple topics in OMBValidationTest.test_max_partitions
Many kafka client libraries (java, librdkafka) perform poorly when
producing to very high partition topics. This leads to one thread not
being fast enough to produce at a reasonable batch size.

In a tier 6 test we were seeing batch size slowly growing all the way to
85K. This isn't great because it means we are not running the test in a
stable state and this becomes more a client benchmark. Further lower
batch sizes are actually more demanding for RP itself.

To avoid this problem we split the total partition count up across
multiple topics and limit each to a max of 5000. Each producer will only
produce to one topic and hence it's easier on them.

This is similar to a pattern that we have already seen employed by
customers with similar high partition counts.

With this change batch size stays stable at 10K from the beginning of
the test. Total producer/consumer count might be slightly different
because of rounding but that shouldn't be an issue as the calculation
for the optimal producer count is fairly conservative already.