c/partition_manager: introduced partition shutdown watchdog timer
Introduced a watchdog that is tracking partition shutdown state. The
watchdog is intended to provide a mechanism that will ease debugging of
partition shutdown issues. The watchdog tracks state of partitions that
were requested to be stopped or removed. When there was no state update
for the time longer than the configurable threshold (by default 30
seconds) the watchdog will emit error log entry which will inform user
about the problem.

Signed-off-by: Michal Maslanka <michal@redpanda.com>