tests: Make the OMB validator use result.json
The OMB validator was not validating p999 even though it was specified
in the validator requirements.

There was two reasons for that:
 - We were using the data from the output of the `generate_charts`
   command which is incomplete and doesn't return p999 e2e latency
 - This wasn't fatal because the validator logic silently failed to
   validate requested validation metrics if they were not in the metrics
   output

This patch fixes both by switching to actually using the `result.json`
file and further invert the validation loop such that it errors out if a
requested validation metric does not exist.