rptest: more typing 2
Type more parts of rptest, concentrating on getting
high_throughput_test to type fully at pyright "basic" setting.

After this change it types fully with some additional ducktape
stubbing (not part of this change).

This change is largely pure typing (which can't affect runtime) though
there are a handful of small runtime changes to assist in the typing,
e.g., asserts that properties are not None immediately prior to using
them which would otherwise by a typing failure. This only changes a
crash at use to a crash immediately before at the assert.