k/replicated_partition: fixed fetch offset validation for follower fetching
When consumer is requesting a fetch from the broker being a follower for
given replica the fetch offset validation logic changes comparing to the
validation on a leader. The out of range error should only be returned
if an offset is not available locally and it is not known to be
committed. That was not the case with the current code. The current code
returned out of range error every time the offset was not committed on
leader. This lead to many not necessary errors returned and resulted in
consumer offsets being reset.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit fb48e0a8f9353abfad80db2dee00a9b2eed07188)