rpc/transport: shutdown connection on unexpected exception
If we happen to end up in this exception handler, chances are that we
never inserted an entry into _requests_queue. This means that there is a
hole in the sequence numbers sequence and the connection is not usable
anymore. Better to reset it.

(cherry picked from commit c517c9b0de68959c9098ee547aeb678b2ab853b8)