rpc: remove netbuf header init check
To check that netbuf header is initialized, we checked correlation_id,
treating 0 as special invalid value. But it turns out, we can get 0 if
correlation_id overflows uint32_t (this can happen on a reasonably busy
connection). The overflow itself is benign as correlation_id just wraps
around to 0 (and it is unlikely that there still remain entries with
these values in the _correlations map) but the check throws. Since it is
just a precautionary check that doesn't add much value, remove it.