archival: Delete replaced segments during spillover GC
Previously, if the spillover was enable we weren running normal GC
during the housekeeping run. Normal housekeeping is responsible for
removing segments evicted due to retention (which is not happening if
spillover region is not empty). But it also removes replaced segments.
So essentially, we have to run both types of GC on every spillover
housekeeping run.