transform-sdk/rust: borrow output record
We only need to borrow the output record. To do that we make the writer
take `Into<BorrowedRecord<'_>>`, so that it can take any type of record
for transforms (we can now do no copy transforms!).

Since `Into<_>` is a trait, we can't have a monomorphic trait take a
monomorphic trait as a function argument (see
https://doc.rust-lang.org/reference/items/traits.html#object-safety). So
we wrap our current trait in a struct that handles resolving the trait
argument to the concrete type our write function will take.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>