rptest: introduce RedpandaCloudTest, refactor
This change changes a bit the inheritance structure of the
RedpandaService tree, and introduces the RedpandaCloudTest.

On the service side, we split RedpandaServiceCloud out so that it is
not a subclass of Service (as it does no use Service nodes or have
a typical Service lifecycle). It shares some common interface with
RedpandaService but this is currently much smaller than the existing
RedpandaServiceBase.

RedpandaCloudTest is the base class for tests which run in the cloud,
analagous to RedpandaTest for tests which run in vanilla environments.
It automatically creates a RedpandaServiceCloud as the .redpanda
property, giving you type safe access to only the methods which make
sense in the cloud.

Issue redpanda-data/core-internal#1002.