admin: added broker_uuids rest api
Added REST admin API allowing user to query for the mapping between the
node_id and UUID. This will allow us to quickly determine a current UUID
and `model::node_id` relation. This may be helpful when dealing with
incidents in which multiple nodes join the cluster with empty data
directories.

Signed-off-by: Michal Maslanka <michal@redpanda.com>