services/redpanda.py: set_cluster_config fix
a previoud change interpreted a {"name": None} meaning that "name"
should be removed in the properties. but this does not work for optional
properties, they will revert to the default value.

the fix is to just pass it in the update field, the property will be
correctly set to nullopt