rm_stm: remove mem_state::last_end_tx
It looks like last_end_tx was added to fix a violation when the code
had lazy catchup on an abort transaction.

This field was added in commit [1] when the abort code looked like this
[2]

```
    if (_mem_state.last_end_tx < r.value().last_offset) {
        _mem_state.last_end_tx = r.value().last_offset;
    }

    // don't need to wait for apply because tx is already aborted on the
    // coordinator level - nothing can go wrong
    co_return tx_errc::none;
}
```

This sort of seemed like an optimization where we could skip the wait
and defer it to next command if needed.

However the code changed later in this commit [3] where an explicit
wait after abort added thus obviating the need for this field.

Just looking at the code anywhere this field is updated, we immediately
have a wait for state machine to catchup until that offset which means
any subsequent wait for it will anyway be a no-op, hence removing it.

This is a part of clean up before moving fields to producer_state.

[1] https://github.com/redpanda-data/redpanda/commit/51f42c1ff6574a664f13f477be489387180c7c57
[2] https://github.com/rystsov/redpanda/blob/51f42c1ff6574a664f13f477be489387180c7c57/src/v/cluster/rm_stm.cc#L656
[3] https://github.com/redpanda-data/redpanda/commit/d8c52549e4c9870cd65e09a9f40e7779db95d924