rm_stm: remove log_state::prepared
preparing/prepared was a phase pre-GA transactions ( ~ 1.5y ago). It is
extremely unlikely that a transaction was stuck in prepared for this
long. It is extremely unlikely that the batches exist in the log for
this long. With these assumptions, cleaning up the related code.

logs a warning for debugging if the state machine encounters such a
batch.