test: Support setting scheduling group in stress fiber
The `cpu_stress_injection_test` test was occasionally flaky. This was
because the stress fiber runs in the admin scheduling group (as that is
where it's started from). This however starves the admin server itself
and hence it can time out sending the actual API response.

To work around that issue we move the stress fibers to run in the main
scheduling group when invoked from the admin api.

To do that we extend the stress fiber api such that the scheduling group
can be specified.

Fixes https://github.com/redpanda-data/redpanda/issues/13701