k/topics: do not use a vector to wait for leaders
Previously we used a simple `std::vector` of futures to make
waiting for partition leaders concurrent. Using a vector has a drawback
when dealing with large number of topics and partitions since it may be
required to allocate large contiguous chunk of memory for a future
vector. In this particular case we may not use a fragmented vector or
chunked fifo as the `when_all` uses a plain vector internally.

To make sure no large chunk of memory is allocated to wait for the
partition leaders changed the logic to use
`seastar::max_concurrent_for_each`.

Fixes: #15908

Signed-off-by: Michal Maslanka <michal@redpanda.com>