wasm: support deploy time validation
Support validating a wasm module without the expensive of compiling and
running it. The validation is mostly a smoke test of the different
things that could go wrong. It's still possible to have a module that
does something "wrong" from an ABI prospective. I think this is good
enough because at some point you have to run the module to make sure it
does the right thing.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>