tx/compaction: don't deduplicate control batches
Control batches (tx_commit, tx_abort) are special data batches that
indicate the completion of a transaction. Because these are technically
data batches, they participate in deduplication.

It was previously thought that as long as we kept the last control batch
in a segment, that deduplicating such records was safe. However,
consider the following sequence:

node 1:
1. begin tx 1, write data batch in tx 1
2. write many non-tx data batches, up to offset 100
3. abort tx 1
4. begin tx 2, write data batch in tx 2, abort tx 2 which is offset 110
5. begin recovery, replicating data to node 2, sending records [0, 50]
6. roll segment

node 2:
7. receives data [0, 50], including [begin tx 1, data tx 1, data]

node 1:
8. self compact segment: since all abort records have the same key, only
   abort tx 2 is kept
9. continue recovery, replicating data to node 2, sending records in the
   range [50, 110], but is missing abort tx 1

node 2:
10. receives data [50, 110], and now has [begin tx 1, data tx 1, data,
    begin tx 2, data tx 2, abort tx 2]
11. roll segment
12. self compact segment: rm_stm on this node doesn't know that tx 1 is
    considered aborted
13. data tx 1 is left in the segment as committed

If instead, at step 8, we had kept [abort tx 1] as well, this wouldn't
have been an issue: node 2 would have replayed the abort and correctly
compacted away data tx 1.

This commit does this disallowing the deduplication of control batches
entirely. These batches are only used in transactions, and are expected
to be uncommon. This happens in the `is_compactible()` method, which is
used at compaction time to determine whether the entire batch should be
kept without consulting further record-based maps.

I'm explicilty doing it here rather than e.g. filtering control batches
when writing to the compaction index because we'll ultimately need to
filter here anyway: older versions of Redpanda will have these control
records in their compaction index.