raft: Improve handling of concurrent truncation
The 'flush_log' method checks the dirty offset before and after flush.
If the truncation happened concurrently it detects this by comparing
offsets (the offset after flush is smaller than the offset before the
flush in this case). Then it uses assertion to check if the flushed
offset is not smaller than the committed offset.

This may not work as expected in case if there is a concurrent
truncation and append. For instance, before the truncation the dirty
offset is 32. The 'flush_log' method records flushed offset 32 and calls
log->flush. Before 'flush_log' is resumed the log gets truncated and
then new value gets added. So the committed offset of the log is now 31
but dirty offset is 32. Because of that the method won't be able to
detect the concurrent flush operation will trigger an assertion.

This commit fixes this problem by checking not only dirty offset but
also a truncation counter. If the counter was incremented then the
log was truncated and new data was added to the end of the log.