storage/tests: remove default reader offset bound
I plan to extend the test suite to cover write caching functionality
too. In the case of write caching we will want to read past the
committed offset.

I couldn't infer a good motivation for bounding test readers to
committed offset. I believe that not having the upper bound by default
can help uncover hidden bugs both in the implementation and in the
testing code.

If anything needs to bound reads, it must do it explicitly like the
regular application code does.