topics_frontend: Add optional fields to topic_result
After creating a topic, we would like to  surface the partition count
and replication factor of the created topic to calling code.

We'll read these values from the topic table and return them if either
  - the topic already exist
  - the topic was created successfully

In all other cases, partition_count and replication_factor are simply
left out of the topic_result.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>