cloud_storage: Use watchding to detect stuck eviction
The logic of the eviction loop did not change. The watchdog callback
only logs error message on error level so it'd be easier to detect this
situation. The alert could be created based on this log message.

(cherry picked from commit c1be88695d99c610276646c76ad5ef36a249b631)