xform/logging/io: Add client::initialize and update log_manager
To create the transform logs topic on demand. Also adds code in log_manager.cc
to perform this operation lazily at flush time.

Additionally, update logging::client to return transform::logging::errc
(or a result<> thereof). This becomes increasingly useful for error
reporting as we start hooking things up to the rest of the cluster.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>