rptest: tolerate missing topic metrics after start
During the recovery test in `test_prefix_truncate_recovery`, a node is
restarted, and metrics are queried to wait until no partitions are
under-replicated.

The test occasionally fails at `assert len(metric.samples) ==
len(nodes)` due to a missing metric for the test topic on the
just-restarted node.

The redpanda server allows metrics to be queried very early in the
server startup flow but the controller log which instructs the server
which partition topics it manages is applied asynchronously.

To resolve this issue, we remove the assertion and instead rely on the
fact that other servers will correctly report under-replicated
partitions using `all(map(lambda s: s.value == 0, metric.samples))`.

Thus, the assert statement is both incorrect and redundant.

An extra check is added to ensure that metrics for all partitions under
test are present and the expression is evaluated for all of them. This
could happen due to a bug or all nodes containing replicas being
restarted and not producing metrics yet. A situation unlikely to happen
in this test but makes the `fully_replicated` method more general and
future proof.

Fixes #16115

(cherry picked from commit 586584f11f6eb3266b340152e5d217ca15f99292)