cloud_storage: Store model::ntp inside remote_partition
Previously, the remote_partition used async_manifest_view instance to
acquire the ntp. This is not safe in some cases because the
async_manifest_view could be disposed. Normally this isn't happening but
this could happen in tests and it might happen in Redpanda when
something is stuck during shutdown process.