r/consensus: do not allow follower to truncate log that is consumable
If an offset was already visible a follower must not be allowed to
truncate it as it may lead to a situation in which an offset is visible
and not readable.

Visible batches has the same replication guarantees as committed batches
as leader still waits for the majority to acknowledge message at given
offset before making it visible to the readers. This makes it possible
not to truncate offsets which were previously visible.

Signed-off-by: Michal Maslanka <michal@redpanda.com>