c/rm_stm: handle cache full error in rm_stm
When the number of allowed virtual clusters maintained in producer
state manager cache is exceeded a cache throws an exception. The
exception is translated to the error code indicating that the cache size
limit was exceeded. The error code is finally translated to kafka policy
violation error. This error is not retryable and will indicate issue
with number of producers.

Signed-off-by: Michal Maslanka <michal@redpanda.com>