rptest: Add feature to allow logs from within test
The log allow list decorator conventionally matches logs on string or
regex. Sometimes finer grained checks are required, where extra context
is needed to see if a log message is acceptable or signals an error.

The AllowLogsOnPredicate class allows a test to register one of its
methods to be called for each error log, that method can then return a
not-None style result to signal that an error is acceptable, and False
or None to signal that an error should cause the test to fail.