rptest: Allow spillover error on anomaly
This test removes a segment from the manifest to generate anomalies. A
side effect of this is that the validation in manifest done during
spillover which checks that the start of the manifest and the start of
spillover must match can fail, because the first segment of the manifest
has been force reset away.

Effectively, spillover manifest is creating by iterating over segment
list in main manifest. This main manifest is missing its first segment
because of the anomaly injection.

When applying the spillover command, the manifest compares the start
offsets of the spillover meta and the main manifest. The start offset
has not changed during the anomaly injection. As a result the
apply-spillover command fails.

This is acceptable because the anomaly injection should cause this
failure, and this change adds the error to allow list.