rpc: Add config flag to enable/disable compression for replies
Add a flag to allow disabling internal rpc reply compression.

Right now we unconditionally enable compression for rpc replies if they
are above the min compression size (1Kib).

In this commit we don't change the default and keep compression on but
we might change that in a follow up commit as the cpu overhead from
compressing/uncompressing larger messages has shown to be detrimental to
overall performance.