cloud_storage: Fix remote_partition issue
The remote_partition_reader can reset to the begining of the partition
in case if it didn't find the segment and both config.start_offset and
config.max_offset are outside of the manifest. When this logic is
applied to the timequery in case if timestamp overshoots last segment in
the manifest we end up having a full partition scan. This commit fixes
this by disabling this logic for timequeries.

(cherry picked from commit f4f67432587d6e7e8ae4369c414462622851e211)