xform/logging: Introduce record_batcher
Encapsulates a storage::record_batch_builder and aims to build up batches
of log records. Batches are kept under a specified maximum size to avoid
data loss. If appending a given kvp would cause a batch to exceed the
specified max, we roll to a new builder, appending the previous batch to
a running collection of completed, size-limited record_batches.

Includes a simple unit test for constructed record_batch sizes.