topic_manifest: move to v::cluster lib
topic_manifest was in the v::cloud_storage library, but it's used mainly
by the v::cluster library, and needs a dependency on
v::cluster::topic_properties type. since v::cloud_storage does not
depend on v::cluster, some proxy object is needed to break this inverse
dependency.

this commit moves topic_manifest to v::cluster, making it possible in
to remove the proxy object.

the position of the file is not changed.