ssx: implement `ssx::sharded_abort_source` using `std::shared_ptr`
The previous implementation of `ssx::sharded_abort_source` had two
potential race conditions;

- If it is stopped then `.local()` is no longer safe to call since
  `.stop()` will delete all local objects. However, other shards do not
have a way of knowing if the abort_source was stopped, hence no way of
knowing if `.local()` is safe to call or not. The new implementation
avoids this issue by using a shared_ptr to delete the local objects
rather than `.stop()`.

- If the parent abort_source aborts then the subscription will make
  cross shard calls to every shard-local abort_source. However, on the
original shard the `ssx::sharded_abort_source` may be stopped and the
shard-local abort_source's deleted before the cross shard calls finish.
Resulting in a use-after-free violation. The new implementation fixes
this in the same way as it did the last issue.