Add erase_to_end to fragmented vector
This is similar to pop_back_n but takes an iterator and is a more
directly replacement for use cases which are already using iterators.

Additionally, allow const_iterator to be converted to iterator, which
is what standard containers do and is important for ergonomics.

(cherry picked from commit 5e5b4e885e5104ae7bc958dc6b88601320946f30)