fragmented_vector: default to dynamic first chunk capacity
In order to simplify the `fragmented_vector` interface we want to remove
the need to explicitly define a chunk size, and also not have small
vectors take too much space. So default the fragment size to be dynamic.

Eventually, we can remove the ability to explicitly state the size of
chunks, and instead always grow to our full allocation size (to reduce
fragmentation).

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>