kafka: rethrow on unknown exceptions in fetch handler
Exceptions from outside the Kafka handler context can bubble up to the
catch present in the handler. This seems to be the way some subsystems
communicate issues with the requests. There is no current listing of
what exceptions subsystems may throw, if/how to recover from these
exceptions, or if a fetch should end as a result of the exception. Hence
for the time being the fetch impl will revert to the behavior of further
bubbling up unknown exceptions to outside of the handler context.

(cherry picked from commit a09d160db0f74034cdf47ae5204afbe5a7218cad)