r/consensus: extend confirmed term semantics to apply for the follower
A confirmed term is used to determine if the state of a replica is up
to date after the leader election. Only after the confirmed term is equal
to the current term one can reason about the Raft group state.

On the leader the confirmed term is updated after first successful
replication of a batch subsequent to a leader election. After the
replication succeed leader is guaranteed to have up to date committed
and visible offsets.

On the follower the confirmed term is updated only when an append
entries request from the current leader may be accepted and follower
may return success.

Signed-off-by: Michal Maslanka <michal@redpanda.com>