c/leaders_table: unify notification mechanism
Using `ntp_callbacks` to wait for the leaders without additional
promises map. When caller requests to wait for a leader we register the
notification which sets the promise value when called. This way we do
not need a separate mechanism to keep track of leadership notifications.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 4eed31fd55411341997ceeb1af9e7a7f5eae5e5f)