cloud_storage: introduce upload object request
The upload object API is used to upload several kinds of small payloads
to cloud storage. This change introduces enum class to lock down the
types of upload which was denoted via strings.

This change also adds a request struct which will be used for aggregate
initialization when calling upload_object for readability. The request
takes optional callbacks to update the remote probe on success, failure
and slowdown. These updates were previously done in hardcoded code paths
for index uploads.