cloud_storage/inventory: Add AWS create-inv action
The create-inventory action for AWS is added. The inventory API is
divided into a low level, vendor specific API and a high level vendor
agnostic API. The high level API will be introduced in the next few
commits.

The low level API hides vendor specific details for common inventory
related operations such as creating an inventory schedule, finding the
latest inventory report, deleting old inventory reports etc. These
actions differ in details across vendors.

The high level API will use a variant to call the correct vendor API
depending on where redpanda is deployed.

This specific pull request focusses on the create-inventory operation
for AWS. This operation schedules an inventory report to be generated on
a given frequency (daily here) and produces reports of a given format
(csv here).

Additionally, a new interface called cloud_storage_api is created which
remote now inherits from. This interface is used in the inventory APIs
mainly to assist in testing. Googlemock can create mocks of this
interface, thus avoiding the need to spin up an imposter to test the
inventory APIs.

While remote now inherits from the new interface and one of its methods
is now virtual, this should not add any overhead for code paths which do
not use the new interface.